sources/pyproject.toml
[project]
name = "koesu-sources"
version = "0.1.0"
description = "Audio source manager for Koesu"
requires-python = ">=3.12"
dependencies = [
    "yt-dlp[default]>=2026.3.17",
    "aiohttp>=3.11.0",
    "aiofiles>=25.1.0",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "protobuf>=6.33.0",
    "uvloop>=0.21.0",
    "orjson>=3.10.0",
    "structlog>=25.1.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.14.0",
    "lxml>=6.0.0",
    "websockets>=14.0",
    "certifi>=2026.0.0",
    "brotli>=1.1.0",
    "mutagen>=1.47.0",
    "ffmpeg-python>=0.2.0",
    "openai>=2.30.0",
    "tavily-python>=0.7.23",
    "ddgs>=9.13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/koesu_sources"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
